rrequirements.txt
gradio>=3.35.2
openai>=0.27.8
python-dotenv>=0.21.0